+++ src/main/java/builderb0y/autocodec/integration/Auto2DFUCodec.java
@@ -1,45 +1,85 @@
 package builderb0y.autocodec.integration;
 
+import java.util.Objects;
+import java.util.stream.Stream;
+
+import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
-import org.jetbrains.annotations.NotNull;
+import com.mojang.serialization.DataResult;
+import com.mojang.serialization.DynamicOps;
+import com.mojang.serialization.MapLike;
 
 import builderb0y.autocodec.AutoCodec;
 import builderb0y.autocodec.coders.AutoCoder;
-import builderb0y.autocodec.decoders.AutoDecoder;
-import builderb0y.autocodec.encoders.AutoEncoder;
+import builderb0y.autocodec.decoders.DecodeException;
+import builderb0y.autocodec.encoders.EncodeException;
+import builderb0y.autocodec.util.DFUVersions;
 
-public interface Auto2DFUCodec<T_Decoded> extends Auto2DFUEncoder<T_Decoded>, Auto2DFUDecoder<T_Decoded>, Codec<T_Decoded> {
+public record Auto2DFUCodec<T_Decoded>(
+	AutoCodec autoCodec,
+	AutoCoder<T_Decoded> coder
+)
+implements Codec<T_Decoded> {
 
-	public static <T_Decoded> @NotNull Auto2DFUCodec<T_Decoded> of(@NotNull AutoCodec autoCodec, @NotNull AutoEncoder<T_Decoded> encoder, @NotNull AutoDecoder<T_Decoded> decoder) {
-		return new Auto2DFUCodec<>() {
-
-			@Override
-			public @NotNull AutoCodec autoCodec() {
-				return autoCodec;
-			}
-
-			@Override
-			public @NotNull AutoEncoder<T_Decoded> encoder() {
-				return encoder;
-			}
-
-			@Override
-			public @NotNull AutoDecoder<T_Decoded> decoder() {
-				return decoder;
-			}
-
-			@Override
-			public String toString() {
-				return (
-					encoder == decoder
-					? "Auto2DFUCodec: { coder: " + encoder + " }"
-					: "Auto2DFUCodec: { encoder: " + encoder + ", decoder: " + decoder + " }"
-				);
-			}
-		};
+	@Override
+	public <T_Encoded> DataResult<Pair<T_Decoded, T_Encoded>> decode(DynamicOps<T_Encoded> ops, T_Encoded input) {
+		try {
+			return DFUVersions.createSuccessDataResult(
+				Pair.of(
+					this.autoCodec().decode(this.coder, input, ops),
+					ops.empty()
+				)
+			);
+		}
+		catch (DecodeException exception) {
+			return DFUVersions.createErrorDataResult(exception::toString);
+		}
 	}
 
-	public static <T_Decoded> @NotNull Auto2DFUCodec<T_Decoded> of(@NotNull AutoCodec autoCodec, @NotNull AutoCoder<T_Decoded> both) {
-		return of(autoCodec, both, both);
+	@Override
+	public <T_Encoded> DataResult<T_Decoded> parse(DynamicOps<T_Encoded> ops, T_Encoded input) {
+		try {
+			return DFUVersions.createSuccessDataResult(
+				this.autoCodec().decode(this.coder, input, ops)
+			);
+		}
+		catch (DecodeException exception) {
+			return DFUVersions.createErrorDataResult(exception::toString);
+		}
+	}
+
+	@Override
+	public <T_Encoded> DataResult<T_Encoded> encode(T_Decoded input, DynamicOps<T_Encoded> ops, T_Encoded prefix) {
+		try {
+			T_Encoded result = this.autoCodec().encode(this.coder, input, ops);
+			if (Objects.equals(result, ops.empty())) {
+				return DFUVersions.createSuccessDataResult(result);
+			}
+			T_Encoded merged = DFUVersions.getResult(ops.mergeToPrimitive(prefix, result));
+			if (merged == null) {
+				MapLike<T_Encoded> map = DFUVersions.getResult(ops.getMap(result));
+				if (map != null) {
+					merged = DFUVersions.getResult(ops.mergeToMap(prefix, map));
+				}
+			}
+			if (merged == null) {
+				Stream<T_Encoded> stream = DFUVersions.getResult(ops.getStream(result));
+				if (stream != null) {
+					merged = DFUVersions.getResult(ops.mergeToList(prefix, stream.toList()));
+				}
+			}
+			if (merged == null) {
+				merged = result;
+			}
+			return DFUVersions.createSuccessDataResult(merged);
+		}
+		catch (EncodeException exception) {
+			return DFUVersions.createErrorDataResult(exception::toString);
+		}
+	}
+
+	@Override
+	public <T_Encoded> DataResult<T_Encoded> encodeStart(DynamicOps<T_Encoded> ops, T_Decoded input) {
+		return DFUVersions.createSuccessDataResult(this.autoCodec().encode(this.coder, input, ops));
 	}
 }